# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PicoClaw Serverless API
  version: 1.0.0
  description: |
    HTTP API for PicoClaw (serverless-oriented runtime).

    Notes:
    - All endpoints except `/health` require Bearer token authentication.
    - `POST /chat` supports both JSON response mode and SSE streaming mode.
    - Memory and conversation history are persisted via mounted storage and reused across separate requests.
    - Scheduled tasks are externally triggered by calling `/task/check`.
    - `POST /control/stop` can be used to request graceful save-and-exit.
    - Every response includes an `X-Request-ID` header for log correlation. If the caller sends an `X-Request-ID` header, the same value is echoed back; otherwise a new `req-<UUID>` is generated.
servers:
  - url: http://localhost:9000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /chat:
    post:
      tags:
        - Chat
      summary: Start or continue a conversation
      description: |
        Sends a user message to the agent.

        Behavior:
        - Without `conversation_id`: creates a new conversation.
        - With `conversation_id`: resumes an existing conversation.
        - With `stream=true`: returns `text/event-stream` (SSE).
      operationId: postChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              newConversation:
                value:
                  message: 请只回答数字：1+1 等于几？
                  sender: verify
                  sender_name: Verifier
                  stream: false
              continueConversation:
                value:
                  message: 上一轮你回答了什么？只回答内容。
                  conversation_id: conv-0df6aee0-86e1-45ac-b33b-c13086b72c36
                  sender: verify
                  sender_name: Verifier
                  stream: false
              streamMode:
                value:
                  message: 逐步思考后给结论
                  sender: verify
                  sender_name: Verifier
                  stream: true
                  max_execution_ms: 120000
      responses:
        '200':
          description: Chat processed successfully (JSON or SSE stream)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                success:
                  value:
                    status: success
                    conversation_id: conv-0df6aee0-86e1-45ac-b33b-c13086b72c36
                    message_id: msg-3aaf5d3f-a17b-4e8e-bec3-a33df4c9af86
                    result: '2'
                    session_id: 3e49773a-b4a6-4c53-8cd9-b3f96c5377e3
                    duration_ms: 6701
                    error: null
                    outbound_messages: []
                    session_end_marker: '[[PICOCLAW_SESSION_END]]'
                    session_end_marker_detected: false
                timeout:
                  value:
                    status: timeout
                    conversation_id: conv-aaa
                    message_id: msg-bbb
                    result: 已生成部分结果
                    session_id: session-xxx
                    duration_ms: 120000
                    error: Execution aborted after 120000ms. Use conversation_id to continue.
                    outbound_messages: []
                    session_end_marker: '[[PICOCLAW_SESSION_END]]'
                    session_end_marker_detected: false
            text/event-stream:
              schema:
                type: string
              examples:
                sse:
                  value: |
                    event: start
                    data: {"conversation_id":"conv-123","message_id":"msg-123"}

                    event: chunk
                    data: {"text":"部分输出"}

                    event: done
                    data: {"status":"success","conversation_id":"conv-123","message_id":"msg-123","result":"最终结果","session_id":"session-123","duration_ms":3210,"error":null,"outbound_messages":[],"session_end_marker":"[[PICOCLAW_SESSION_END]]","session_end_marker_detected":true}
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                missingMessage:
                  value:
                    error: message is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation ID not found when resuming
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                notFound:
                  value:
                    error: 'conversation_id not found: conv-missing'
        '409':
          description: Conversation is currently processing another request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationBusyError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatInternalErrorResponse'
    get:
      tags:
        - Chat
      summary: List all conversations
      operationId: listConversations
      responses:
        '200':
          description: Conversation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/{conversation_id}:
    get:
      tags:
        - Chat
      summary: Get conversation state
      operationId: getConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Conversation state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationStatusResponse'
        '400':
          description: Missing or invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Chat
      summary: Delete a conversation
      description: |
        Deletes a conversation and all related data (messages, outbound messages, tasks).
        Returns 409 if the conversation is currently running.
      operationId: deleteConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '204':
          description: Conversation deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conversation is currently running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationBusyError'
  /chat/{conversation_id}/messages:
    get:
      tags:
        - Chat
      summary: Get conversation messages
      operationId: getConversationMessages
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationMessagesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_TOKEN
  parameters:
    ConversationId:
      name: conversation_id
      in: path
      required: true
      description: Conversation ID
      schema:
        type: string
        minLength: 1
...
  headers:
    X-Request-ID:
      description: |
        Request tracking identifier for log correlation.
        Echoes the caller-provided `X-Request-ID` header, or a generated `req-<UUID>` if none was sent.
        Present on every response, including error and unauthenticated responses.
      schema:
        type: string
        example: req-a1b2c3d4-e5f6-7890-abcd-ef1234567890
  responses:
    Unauthorized:
      description: Unauthorized (missing or invalid bearer token)
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorized:
              value:
                error: Unauthorized
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
...
    ChatRequest:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
        conversation_id:
          type: string
          description: Existing conversation ID for multi-turn continuation.
        sender:
          type: string
          description: Sender identifier.
          default: user
        sender_name:
          type: string
          description: Sender display name.
        stream:
          type: boolean
          default: false
          description: Return SSE stream when true.
        max_execution_ms:
          type: integer
          minimum: 1
          description: |
            Request-level timeout in milliseconds.
            Effective timeout is min(max_execution_ms, env MAX_EXECUTION_MS).
    OutboundMessage:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - text
        - sender
        - created_at
      properties:
        id:
          type: integer
        conversation_id:
          type: string
        text:
          type: string
        sender:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ChatResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - conversation_id
        - message_id
        - result
        - duration_ms
        - outbound_messages
        - session_end_marker
        - session_end_marker_detected
      properties:
        status:
          type: string
          enum:
            - success
            - timeout
            - error
        conversation_id:
          type: string
        message_id:
          type: string
        result:
          type: string
          nullable: true
        session_id:
          type: string
          nullable: true
        duration_ms:
          type: integer
          minimum: 0
        error:
          type: string
          nullable: true
        outbound_messages:
          type: array
          items:
            $ref: '#/components/schemas/OutboundMessage'
        session_end_marker:
          type: string
          description: Marker string that indicates the assistant declared session completion.
        session_end_marker_detected:
          type: boolean
          description: True when the marker is found in result or outbound messages.
    ChatInternalErrorResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - conversation_id
        - message_id
        - result
        - error
      properties:
        status:
          type: string
          enum:
            - error
        conversation_id:
          type: string
        message_id:
          type: string
        result:
          type: string
          nullable: true
        error:
          type: string
    ConversationStatusResponse:
      type: object
      additionalProperties: false
      required:
        - conversation_id
        - message_count
        - last_activity
        - status
      properties:
        conversation_id:
          type: string
        session_id:
          type: string
          nullable: true
        message_count:
          type: integer
          minimum: 0
        last_activity:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - idle
            - running
...
    ConversationListResponse:
      type: object
      additionalProperties: false
      required:
        - conversations
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationStatusResponse'
    ConversationMessage:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - role
        - content
        - created_at
      properties:
        id:
          type: string
        conversation_id:
          type: string
        role:
          type: string
          enum:
            - user
            - assistant
            - system
        sender:
          type: string
          nullable: true
        sender_name:
          type: string
          nullable: true
        content:
          type: string
        created_at:
          type: string
          format: date-time
    ConversationMessagesResponse:
      type: object
      additionalProperties: false
      required:
        - conversation_id
        - messages
      properties:
        conversation_id:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMessage'
    ConversationBusyError:
      type: object
      additionalProperties: false
      required:
        - error
        - conversation_id
      properties:
        error:
          type: string
        conversation_id:
          type: string
...